Add files via upload
In this Dashboard, I have shown various insights such as the Sum of amount by customer name, the Sum of quantity by category, the sum of profit by state, profit by month, the sum of profit by sub-category in each quarter could be checked by selecting different states. 

I am unable to share the link from MS Power BI as it's asking for an official account or a student account.